// src/interfaces/liquid-vapour-loads.interface.ts
export interface ILiquidVapourLoad {
  position: number | string;
  liquidTemp: number | string;
  vapourTemp: number | string;
  liquidMassFlow: number | string;
  vapourMassFlow: number | string;
  liquidVolFlow: number | string;
  vapourVolFlow: number | string;
  liquidMolWeight: number | string;
  vapourMolWeight: number | string;
  liquidMassDens: number | string;
  vapourMassDens: number | string;
  liquidViscosity: number | string;
  vapourViscosity: number | string;
  surfaceTension: number | string;
}